+++ packages/moengage_flutter/moengage_flutter_platform_interface/lib/src/internal/callback/callback_cache.dart
@@ -16,4 +16,7 @@ class CallbackCache {
 
   /// InApp Dismiss Callback
   InAppDismissedCallbackHandler? inAppDismissedCallbackHandler;
+
+  /// Logout Complete Callback
+  LogoutCompleteCallbackHandler? logoutCompleteCallbackHandler;
 }